backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
pydantic-settings
firebase-admin
pymongo
httpx
